@@ -7,8 +7,8 @@
   "repository": "https://github.com/Pinelab-studio/pinelab-vendure-plugins",
   "license": "MIT",
   "private": false,
-  "main": "dist/vendure-plugin-invoices/src/index.js",
-  "types": "dist/vendure-plugin-invoices/src/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "files": [
     "dist",
     "README.md"
@@ -16,8 +16,8 @@
   "scripts": {
     "start": "yarn ts-node test/dev-server.ts",
     "serve": "nodemon --watch \"src/**\" --ext \"ts,json\" --exec \"ts-node test/dev-server.ts\"",
-    "build": "rimraf dist && yarn graphql-codegen && tsc && copyfiles -u 1 'src/ui/**/*' dist/vendure-plugin-invoices/src",
-    "test": "jest --preset=\"ts-jest\""
+    "build": "rimraf dist && yarn graphql-codegen && tsc && copyfiles -u 1 'src/ui/**/*' dist/",
+    "test": "vitest run"
   },
   "dependencies": {
     "adm-zip": "^0.5.9",
@@ -30,21 +30,19 @@
     "@graphql-codegen/typescript": "^2.4.8",
     "@graphql-codegen/typescript-operations": "^2.3.5",
     "@types/adm-zip": "^0.4.34",
-    "@types/jest": "^27.4.1",
+    "@swc/core": "^1.3.59",
+    "unplugin-swc": "^1.3.2",
     "@types/tmp": "^0.2.3",
-    "@vendure/admin-ui-plugin": "1.9.5",
-    "@vendure/core": "1.9.5",
-    "@vendure/testing": "1.9.5",
-    "@vendure/ui-devkit": "1.9.5",
+    "@vendure/admin-ui-plugin": "2.0.4",
+    "@vendure/core": "2.0.4",
+    "@vendure/testing": "2.0.4",
+    "@vendure/ui-devkit": "2.0.4",
     "aws-sdk": "^2.1099.0",
     "copyfiles": "^2.4.1",
-    "jest": "^27.2.0",
-    "localtunnel": "^2.0.1",
-    "nock": "^13.1.3",
     "nodemon": "^2.0.15",
-    "ts-jest": "^27.0.5",
-    "ts-node": "^10.7.0",
-    "typescript": "4.3.5"
+    "ts-node": "^10.9.1",
+    "typescript": "4.9.5",
+    "vitest": "^0.30.1"
   },
   "gitHead": "89507cfaa85a9fe1fb204f382f6ba730138fcc27"
 }
