@@ -17,13 +17,13 @@ import { DataStrategy } from './api/strategies/data-strategy';
 import { DefaultDataStrategy } from './api/strategies/default-data-strategy';
 import { LocalFileStrategy } from './api/strategies/local-file-strategy';
 import { loggerCtx, PLUGIN_INIT_OPTIONS, PLUGIN_NAME } from './constants';
-import { invoicePermission, StorageStrategy } from './index';
-import { logIfInvalidLicense } from '../../util/src/license';
+import { StorageStrategy } from './index';
+import { invoicePermission } from './api/invoice.resolver';
 
 export interface InvoicePluginConfig {
   /**
-   * Licensekey for commercial use of this plugin.
-   * For more information visit https://pinelab-plugins.com/plugin/vendure-plugin-invoices/
+   * @deprecated We are moving this paid plugin to the Vendure Marketplace soon, so a licensekey won't be needed anymore.
+   * Existing customers will be migrated to the new system ofcourse.
    */
   licenseKey?: string;
   /**
@@ -44,10 +44,11 @@ export interface InvoicePluginConfig {
   ],
   controllers: [InvoiceController],
   adminApiExtensions: {
-    schema,
+    schema: schema as any,
     resolvers: [InvoiceResolver],
   },
-  configuration: (config) => InvoicePlugin.configure(config),
+  configuration: (config: RuntimeVendureConfig) =>
+    InvoicePlugin.configure(config),
 })
 export class InvoicePlugin {
   static config: InvoicePluginConfig;
@@ -55,7 +56,6 @@ export class InvoicePlugin {
   static init(
     config: Partial<InvoicePluginConfig> & { vendureHost: string }
   ): Type<InvoicePlugin> {
-    logIfInvalidLicense(Logger, PLUGIN_NAME, loggerCtx, config.licenseKey);
     InvoicePlugin.config = {
       ...config,
       storageStrategy: config.storageStrategy || new LocalFileStrategy(),
@@ -64,7 +64,9 @@ export class InvoicePlugin {
     return this;
   }
 
-  static async configure(config: RuntimeVendureConfig) {
+  static async configure(
+    config: RuntimeVendureConfig
+  ): Promise<RuntimeVendureConfig> {
     config.authOptions.customPermissions.push(invoicePermission);
     if (this.config.storageStrategy) {
       await this.config.storageStrategy.init();
